Create a preprocessor for highlighting prism code
The processor implements a prismProcessor object, this object test the
component, it must be of type element and represent a code tag, its
parent must be a pre tag and contain a class that starts with language-

If this element passes all the texts then it we will modify its
component property using the Code function in code.js. This function is
in charge of calling the highlight method, to apply the token classes to
the code. It will return the tokenized code inside a code with its
respective class.