Add code styles for every pre tag
Since frontity uses SSR, the processor doesn't behave the same way as it
would in a normal wordpress installation. The pre code class is not
automatically filled. Therefore we use css to apply a "language-x"
styles to every precomponent. It would be a good idea to change this
behaviour in the future and change the prism processor instead.