Prevent undesired enlargament of images in pages
The image preprocessor in index.js cause that, when an image is smaller
than the width, any image becomes enlarged to fit the maximum size,
which causes some blurry images. However the component is useful, so css
changes were made. Since the css is inside the package there was no
remedy but to add a !important tag to container and img css in post.js