Change Gr icons for Ionicons in social component
Apparentely React-icons doesn't support tree shaking out of the box, or
at least in not all of its packages, and it has been a problem for some
time. You can read more about the issue here:
https://github.com/gatsbyjs/gatsby/issues/27276
I think the problem is inherent to Gr icons so I changed the Gr icons
for Ionicons, so icons don't impact in the Bundle server size.