Search bar overlay now handles enter submit
The previous version of the search bar only allowed a search to be made
when the search button was clicked, because of an issue in the way the
toggleSearchBar action handled the event. Now the search bar works with
any type of submit event, whether it be a click or the enter key. The
overlay div closes on a successful search.