Add a close button to newsletter modal
The newsletter modal has a close icon at the top right. However in
certain sizes, specially in cellphones the close mark is outside the
screen. Therefore a close button next to subscribe me button was added.
Also the preventDefault() function was added to the event so the
newsletter form doesn't redirect to mailchimp when the close button is
clicked. The time to show the modal has been increased to 60 seconds for
experimentation purposes.