Enable search by relevance via orderby param
Frontity doesn't enable search by relevance by default at today's date. Although search
parameter is used, no orderby parameter is established in the default
handler. That causes the following problems:
- Adding orderby to state.source.params brings problems when routes
  other than search are accesed
- Other option is to modify the state between each fetch to search
  endpoints, but it leaves us with a messy code
Therefore a extra handler was added, it handles all the routes that
contain "s" parameter and it fetch the resources with the orderby
parameter set to "relevance". Since the code is almost the same as the original (only minor changes) the code was taken blatantly from:
https://github.com/frontity/frontity/blob/dev/packages/wp-source/src/libraries/handlers/postTypeArchive.ts
and added to the init function inside the frontity state.theme