Enable cookies selection in the consent banner
A simple checkbox was added, this checkbox decides whether or not the
google analytics consent will be added if clicked on the consent button.
The function to be executed is established depending on the state of the
cookie list menu. If the cookie list is opened, the user can select
which cookies he/her will accept, otherwise all cookies will be
accepted.

Also, the type search bar type has been changed to search