Place a newsletter signup form after every post
The mailchimp form component is being reused conditionally changing the
css whether there is a prop called formTitle, that passes a text to be
used as the form title. A prop called descriptionTitle does the same as
formTitle but doesn't render css conditionally.
This commit also eliminate props being passed from post component so
they can be obtained directly from the state.