Add support for category tags
Category tags were placed before reading time tag in list of posts and
single post components. The tag redirects to their respective category.
Pagination was also modified to be compatible with categories. The
router.link property named route is used in the link construction.
Basic CSS for tags is added.