Modify similar post component
The previous similar post component had a very messy code, with lots of
steps to generate a pseudo random selection of posts based on the parent
category, this new code is way more cleaner, because all the post
generation was moved to the backend and is in charge of the YARPP
plugin. A custom endpoint was added with its respective handler.