Delete title props in FontAwesomeIcon Component
FontAwesome has a bug that creates a random ID when a title is passed as
a prop to the FontAwesomeIcon component. The solution is to wrap the
component inside a span and apply the title to the span tag, or not
using a title at all. All titles were deleted except the one from the
search icon.

Also, add pointer cursor to interactive components.